fix overflow

+++ handlers/handler.go
@@ -26,7 +26,7 @@ func Fibonacci(c *gin.Context) {
 
 	res, ok := cache.Get(num)
 	if !ok {
-		res = utils.FibNumbers(int(n))
+		res = utils.FibNumbers(int64(n))
 		cache.Set(num, res)
 	}
 	c.JSON(200, gin.H{
